Have all opcode analysis operate on MethodTransform
Summary:
Right now we convert our bytecode between a vector of DexInstructions* and an
intrusive list in the form of MethodTransform several times in the course
of running our passes. This is not great for a number of reasons:

* Converting back and forth is slow
* Makes introducing an IR as part of the vector -> list ballooning step
  difficult
* Since only one representation is active at any given time, if you try to
  access the "wrong" representation, redex crashes

This diff therefore changes all our opcode operations (aside from loading /
outputting) to operate on MethodTransform instead. It also stores the
MethodTransform on DexCode itself rather than in a separate cache --
every DexCode now has an active MethodTransform, so there's no need for a
cache.

Much effort was spent in the making of this diff to track down discrepancies in
output before and after the refactor. After fixing a bunch, the following
remain:

1. When a method has large jump offsets, we may emit two opcodes for a single
  jump (e.g. if-eqz + goto) during the sync step. By not running a sync step
  immediately after SimpleInline, and instead running localDCE first, we
  reduced the number of instances where we had to emit two opcodes per jump.
2. Another effect of not running sync immediately after SimpleInline is that we
  produce slightly different CFGs. These are isomorphic for most purposes, but
  ConstantPropagation traverses the graph somewhat oddly and propagates
  a few less constants after this change.[1]
3. The line number map seems to have more entries after this change. It's not
  clear to me what causes this or where the extra entries end up, but the
  entries still seem to map to sensible line numbers.[2]

Since these all seem benign I think we're good to go; but to aid future
debugging by isolating changes in output, I've left in a call to sync_all() at
the end of SimpleInline, which I will remove in a follow-up diff.

== Future Plans ==

* Have MethodTransform replace DexCode as a member of DexMethod. DexCode will
  only exist in the load / output phases
* Rename MethodTransform to something like IRCode

Reviewed By: dariorussi

Differential Revision: D4673232

fbshipit-source-id: 236e491b04596c03a51c3830a2fbadefdc816180