Fix the string length peephole pattern
Summary:
"stringA".length() is replaced by its length in compile time. But we removed `const-string` as well. We found a case where this delete string was used later.

Example:
https://github.com/google/gson/blob/master/gson/src/main/java/com/google/gson/stream/JsonWriter.java

`setIndent` was inlined to a Gson.newJsonWriter(Writer writer). The argument was replaced with "  ". Peephole replaced `"  ".length()` with a const of 2. But we also removed the string "  ". Later in the code, it accesses the string, resulting in verification error due to type mismatch.

We leave the string in this phase. Later passes like DCE should cleanup it.

Reviewed By: justinjhendrick

Differential Revision: D4732424

fbshipit-source-id: fc5015548c20bf183faac0483c2ec057acd8882f