Introduce IRInstruction
Summary:
DexInstruction is a bit too close to the metal for implementing various
bytecode analyses and transformations, so I've made our optimizations operate
on a slightly higher-level IRInstruction.

Right now the only differences between the instructions are:

* IRInstruction can store register IDs of arbitrary size, instead of being
  limited by the encoding details of DexInstruction
* IRInstructions eliminate the "2addr" opcodes. Since it stores the dest
  register in its own field and the src registers in a vector, aliasing
  dest() and src(0) would actually be more complicated. Eliminating them
  simplifies a number of our analyses.

Note that assigning high registers to IRInstruction will still cause redex
to crash in the output phase, as we try to convert them to DexInstructions
naively... I will address that in an upcoming diff.

Reviewed By: newobj

Differential Revision: D4720603

fbshipit-source-id: 3c8a1becbc32f8f4f80fa29a72f0ecf96a48757b