Track instruction that sets the fields
Summary:
As title.
This change only adds for each field, the instruction that sets the value of the field at that point.
To be used in a later diff, since if we are going to use an extra register for the field, we need to know when and what register to initialize it with.

Reviewed By: int3

Differential Revision: D4746781

fbshipit-source-id: 91b412fb45dc73ead3969409ddecf9e4f4655130