Fixed a typo in the WTO algorithm that the unit test failed to expose
Summary: There was a typo in the algorithm for computing the weak topological ordering: the visit method mistakenly returned NodeId instead of uint32_t. Since the unit test instantiates the class with NodeId = int, the class is type-checked after being instantiated and all integer types are silently interconvertible in C++, this very obvious error went unnoticed. Type safety in C++ still very much belongs to the realm of wishful thinking :-(

Differential Revision: D4764630

fbshipit-source-id: bbd2a1db618720a69dcc8c82a54295bdd192fe74