Remove try-catch check in inliner
Summary:
The logic is actually somewhat inaccurate now because the "source of
truth" about which try entries a method has now resides in the FatMethod
and not in the DexTryItem.

While we could fix it, we've enabled inlining of methods w/ try-catch
for some time now, and it seems to work fine, so I think it's safe to
remove the logic that gates this feature.

This paves the way for the DexCode -> IRCode conversion; IRCode will not have
a `get_tries()` method.

Reviewed By: newobj

Differential Revision: D4768775

fbshipit-source-id: efb8ab26734abb47290b3696913c01a7ba1cde43