redex.py should understand TRACEFILE
Summary:
Previously, it was only used by the redex-all binary but not by the
wrapper script.

A bunch of complexity was introduced in order to prevent redex-all
and redex.py from stepping on each others' toes when writing to the
file. (Note that having redex-all open() the file under append mode is not a
 desirable solution as we still want to overwrite the file when we run it
outside the wrapper script.)

The solution was to have redex-all to treat integer-parseable TRACEFILE values as file
descriptors, and have redex.py change TRACEFILE to point to the descriptor of the
log file it had already opened.

Reviewed By: emma0303

Differential Revision: D4771576

fbshipit-source-id: cac3c4263c33ac51a52f6165de52d8b8f7a8cad4