Rewrite "external catch types" check w/o doing get_tries()
Summary:
The logic is actually somewhat inaccurate now because the "source of
truth" about which try entries a method has now resides in the FatMethod
and not in the DexTryItem.

This diff rewrites the check to walk the FatMethod looking for the catch types
instead of relying on the DexTryItem.

This paves the way for the DexCode -> IRCode conversion; IRCode will not have
a `get_tries()` method.

Reviewed By: dariorussi

Differential Revision: D4768840

fbshipit-source-id: 589a5dc862b259c4721eceb65db5f0ea587778bd