Construct IRInstructions without going through DexInstruction
Summary:
This is a cleanup of the refactoring diff {D4720603}.

Previously we had to construct a DexInstruction just to build an
IRInstruction because a number of IRInstruction properties depended on
the value of instance methods on DexInstruction. Those methods really
didn't have to be instance methods, since they just needed the value of
the DexOpcode.

This diff makes those methods global functions instead.

Reviewed By: dariorussi

Differential Revision: D4761022

fbshipit-source-id: 6e208ac325f08b40eb1613b6c7c5cfbdcad18f85