Store MethodTransform directly on DexMethod; call it IRCode
Summary:
... and make sure we can never be in a confusing state where some methods are
represented as DexCodes and others as IRCodes.

== Notes if you're rebasing upon this diff ==

  cd opt
  find . -type f -exec sed -i 's/DexCode/IRCode/g' {} +
  find . -type f -exec sed -i 's/\.get_entries()//g' {} +
  find . -type f -exec sed -i 's/->get_entries()//g' {} +

should get you a decent amount of the way towards getting your changes to
compile.

== Motivation ==

IRCode (née MethodTransform) contains information from a number of DexCode
members: opcodes, try items, and debug items. However, when we converted from
DexCode to IRCode, we only cleared the list of instructions while leaving the
try and debug items intact. Any analyses that traversed those items would
risk looking at a stale view of the code.

Additonally, the only fields that DexCode had and that MethodTransform missed
was the register sizes. Having to pass a DexCode alongside the MT just to
access those fields was awkward.

To solve these problems, the DexCode -> IRCode conversion now fully deletes
the DexCode, and includes the register size fields in IRCode.

It's also nice to have one less indirection when analyzing code.

I'm still not super pleased that we're still able to encounter runtime crashes
when trying to access the DexCode / IRCode at the wrong time, but this is
probably good enough for now...

Reviewed By: minjang, dariorussi

Differential Revision: D4769415

fbshipit-source-id: 73a8950412bb6ab29cce72498c4aa31abe97aea2