When creating a new method in SingleImplPass, make sure it isn't already concrete.
Summary:
When running without proguard, it is possible for a method to be removed
by RemoveUnreachablePass, which is then re-created by SingleImplPass. Since the
original method stays alive inside RedexContext, it still has all the old code,
annotations, etc. This caused a crash.

This is not a good long-term fix.

Reviewed By: int3

Differential Revision: D4790416

fbshipit-source-id: 1611ac88797675ba709dea6236dceab5abb98a1b