conserve output registers in peephole
Summary:
Previously, PeepholeV2 did not preserve the state of the registers
after an optimization. This has caused at least one crash (t16720000).

Many of the optimizations in this file were just waiting to cause another
crash like that one.

To fix this, I write the values into the registers in the replacement
instructions. We rely on Dead Code Elimination (DCE) to clean this up

I've run into a few complications while making this fix:
 * We (usually) only have enough registers to write to the registers after
   the new code. This causes issues when the pattern ends with an invoke.
   Specifically, we cause a VerifyError by placing const-strings (or similar)
   before a move-result-object. I got around this issue by additional
   peephole instructions "*_WithoutMoveResult" and changing the order that
   we apply optimizations (one at a time).
 * The other complication is that when two symbolic registers (like Register::A
   and Register::B) map to the same real register (like v1), you must write
   to them in the same order as before the optimization (else you ovrewrite
   something that would later be used).

See the comments in the diff for more specifics

Reviewed By: minjang

Differential Revision: D4784032

fbshipit-source-id: d2e61041ad1070078f6aed2a611f326fb0c044b9