Remove trivial build method check
Summary:
As title

In order to properly remove that check, we need to check the method can
actually be inlined.

For now, reusing inline functionalities
* make inline_callees public

For the moment, not too much changes, giving that only a few of the builders
had a "non trivial build method" (by our definition) that could be inlined.

Reviewed By: int3

Differential Revision: D4792401

fbshipit-source-id: af524fc8d74037f43d51a66c1ad31fb34bc89ba5