Remove unneeded method and class symbols
Summary:
We currently write three types of symbols per method or class based on:

1. dex ordinal number
2. dex checksum
3. top 4 bytes of dex signature as a little-endian uint32_t

Of these only the last one is actually useful. The rest are flawed in different
ways (1 because of separate dex stores restarting the counting and a large
amount of effort to find the ordinal on device; 2 because dexopt and dex2oat
change the checksum).

Reviewed By: int3

Differential Revision: D4808152

fbshipit-source-id: 645cd057c2f0ad28aa8245d0bda260ac056cc721