Change how we detect all methods were inlined
Summary:
Before, we where checking that the number of inlined methods is the same as the
number of methods I want to inline. This didn't work properly for one
of the test cases, when even though the `build` method was inlined, number of
inlined methods didn't show up correctly. Changed to actually check there are
no more builder methods that where not inlined. (will look into why the other method didn't work)

After this change, the number of builders that where removed also changed
* before

  Classes removed 876
  Methods removed 1026
  Fields removed 1908
  Methods cleared 793

* after

  Classes removed 895
  Methods removed 1064
  Fields removed 1949
  Methods cleared 828

Differential Revision: D4838624

fbshipit-source-id: ef434fbe8778781ae86d04876700aa1655bf0580