Set inliner_config properties same as for the simple inliner
Summary:
As title.

I was setting the fields for the inliner_config, with the default values.
Properly set them with the values from the simple inliner.
TODO: we could add flags for them later.

This change would only expand the methods we accept to inline, to those that
the simple inliner would.

For now, this doesn't change the methods we inline.

Reviewed By: int3

Differential Revision: D4851659

fbshipit-source-id: ba79a447240f1c4ef2da7f262676c95c92ccabd7