Implementation of a monotonic fixpoint iterator with widening.
Summary: Please note that the design of the abstract domain API and fixpoint iterator was driven in large part by performance considerations. The next step is to create a bridge between this generic infrastructure and the Redex internal form. Once this is done, the infrastructure can be incrementally extended with more advanced algorithms and structures (fixpoint iterator with both widening and narrowing, abstract environments, more abstract domains, etc.).

Differential Revision: D4794388

fbshipit-source-id: cd54f311bf28ac544010cd7679e38fcd396c86a4