add config for RedundantMoveElimination
Summary:
RME caused a crash here T17251764 because RME
doesn't keep track of types in the registers,
but the verifier does.

Adding this to the config should make it safe to
run this pass:
```
"RedundantMoveEliminationPass": {
  "eliminate_const_literals": false
}
```

This will prevent the verify errors because const literals
are the only time (I think) that the verifier updates
the type in a register AFTER it has been written to.

* `move vA, vB` is safe because that just copies the type over
* `const-class` and `const-string` are safe because any register loaded with this should have the same RegisterType

See the discussion on D4856953 for more details

Reviewed By: int3

Differential Revision: D4873454

fbshipit-source-id: 2d0611e16fefdbf7ad5e4077ccec68c2633ded8b