Fix handling of files specified in PG command line
Summary:
{D4503071} added a fallback lookup for -libraryjar and -include files that had
buck-out in their name. This should no longer be necessary with thezhangwei's fixes.
Additionally, the implementation of the fallback had a bug -- depending on the
exact path, redex could continue running even if the file was not found,
without alerting the user in any way.

This commit removes the fallback logic.

Reviewed By: thezhangwei

Differential Revision: D4907774

fbshipit-source-id: d45dbe5a603bec227214d2b3380c16f947f27f30