Flatten IRInstruction hierarchy
Summary:
By storing the method/field/string/type refs in a member union.

Also, instead of storing the union's tag as a field, we instead have it as a macro
argument and generate a switch statement that returns the appropriate tag for
a given opcode.

Benefits:

* Eliminates the many of the static_casts we have
* IRInstruction now uses less memory (no vtable, no type tag)
* (Future work) I want to make it possible to deserialize text representations of IRInstruction, and a flat hierarchy simplifies this.
* (Future work) Having the ref tag known at compile time can help us codegen some useful stuff, e.g. a better dex_asm implementation

While at it, I also renamed a number of methods with names that made more sense (to me). In particular:

* `has_types` -> `has_type`. Same for `has_{strings|methods|fields}`. There is only ever one ref per instruction
* `field()` -> `get_field()`. All other ref accessors use the `get_` prefix, so this is more consistent
* `rewrite_foo()` -> `set_foo()`

Reviewed By: justinjhendrick

Differential Revision: D4938057

fbshipit-source-id: f42742b24dbe747c8561954bf8d9a7e606bcdc0e