Change representation of WtoComponent's end iterator
Summary:
{D4937103} didn't quite solve the edge cases fully. Since we were representing
the end iterator of a WtoComponent as a nullptr, we had to use
m_next_component_offset in the start iterator to figure out if we were at the
end the list of components or if we had any subcomponents, and return the end
iterator for those special cases. This complicated the iterator logic and was
buggy -- we ended up conflating the "no more components" and "no more
subcomponents" cases. By representing the end iterator as just another pointer
into the vector, those cases get handled more naturally.

I also added the ability to print the test graphs in the DOT format, and
regenerated the ascii art graph comments by piping that DOT output into
Graph::Easy. Makes for easier visualization / maintenance.

Differential Revision: D4967889

fbshipit-source-id: 3f5272bd72c00e58f09d0e133563648069664a36