Abstract method devirtualization logic to its own reusable class
Summary:
Relocate the logic of devirtualizing methods to a reusable class, MethodDevirtualizer.
The API here takes a scope and a list of target classes.
It only applies devirtualization on the vmethods and dmethods of the target classes.

The idea is that this absraction makes it easier for hot/cold split to apply devirtualization selectively only on the classes we are interested in.

Reviewed By: minjang

Differential Revision: D4975308

fbshipit-source-id: ff0478db31b1d7f943b23e73dd577a267159bdd8