Clean up of type erasure - 1
Summary:
Wei (thezhangwei)
This is not a lot for now.
I have removed the pass to be TypeErasure as it should be and for now it only erases interfaces ;)
I have also changed some formatting, hope you don't mind. Mostly to bring things back to the infamous 80 lines.
I am going to move the interface removal in its own file (still same pass, same opt, same directory) so we can work on the type erasure as a composition of steps.
Need to do some more serious reworking but putting this out there to get your comments if any

Reviewed By: thezhangwei

Differential Revision: D4975079

fbshipit-source-id: fea3429373d5be0ee34832f743f73d56cb1f44d5