Move interface removal out of main file
Summary:
Moving interface removal as a step in type erasure.
I think we are starting to have too many includes, maybe we should put all of the includes into a single one (either in the pass or another include outside the pass).
Also I am going to do another cleanup pass and define the classes we can erase so we can move in parallel in a much easier way. Sorry if I am slowing things down

Reviewed By: thezhangwei

Differential Revision: D4979781

fbshipit-source-id: 77cdb8e38054a853a5b1f18ceda34afb56088a2f