Make a generic conservative verifier that runs at the end of every pass
Summary:
This code checks that we have not references to deleted types or members.
It only makes sense if run at the end of all passes.
It basically runs against the final scope and relies on the fact we do not delete our objects and so it can determine if a reference to what used to be an internal member is left behind.
Running on current master finds 53 dead references and potential crashes as reported in P57316426.
I am not sure about the implication of those as I'd expect those classes to be unverified so don't know yet the impact.
Right now it only prints bad references via TRACE and though it collects them, it does nothing about it.
I am considering making the build fail but I wanted to hear what people think about it.

Reviewed By: justinjhendrick

Differential Revision: D4909189

fbshipit-source-id: 87106f9f62f540567043d3348af6b709fa105297