builders: Fix escape analysis
Summary:
My initial implementation of escape analysis had a bug -- in order to mark the
register containing the `this` argument as tainted, it marked that register
while analyzing the first instruction of the method. However that meant that
the register was only marked as tainted *after* the first instruction. So if
the `this` register was used as a src in for an invoke instruction at the start
of a method, we would not have noticed that it escaped.

Fixing that bug caused all builders to be marked as "escaped" because their
constructors would invoke the parent constructor with the `this` register. So I
also updated the analysis to treat `<init>` methods specially. The net result
is that more builders get removed than before.

(Noticed this while working on my refactor to replace ins_size with special
opcodes)

Reviewed By: emma0303

Differential Revision: D4985061

fbshipit-source-id: 73cea9a333b8280c6d5c9a5ee355564a94f4bc82