cfg: Add API for creating single exit block
Summary:
This is necessary for running backward analyses with the abstract
interpretation framework. If there are multiple exit nodes, the WTO
iterator will fail to iterate over some nodes in the backwards direction.

Note that it's not as simple as looking for BBs with `return` or `throw`
opcodes. Infinite loops are a valid way of terminating dex bytecode too.
As such, we need to find all SCCs without successors.

I used Tarjan's algorithm for this (adapting some of the implementation style
from the WTO code). It's a little messy because I added a boolean value to keep
track of whether any vertex in the current SCC has a successor edge that points
outside itself. I could probably have done it in two passes instead but oh
well...

The single exit node is only created on demand; it's not run as part of
build_cfg() so as not to slow things down unnecessarily. I have not benchmarked
the overhead though.

I also added an `entry_block` method for completeness.

Differential Revision: D5001144

fbshipit-source-id: 590ceadab5e791b66887667dd26e900fdc9954d3