Reject duplicate classes or dexes
Summary:
We found an APK with duplicate dexes. We should reject such APKs.

In order to that, we check whether the same class name is loaded. The check does not increase any overhead; the code is inside a mutex. But DexClass contains one more `std::string` to remember the origin dex location.

Reviewed By: int3

Differential Revision: D4989874

fbshipit-source-id: 0a7c2c9095d70dfa8331f580ef39ddf2dc237643