Make invoke instructions easier to analyze
Summary:
Provide APIs to:

1. Convert /range instructions into non-range form
2. When an invoke instruction has a wide argument, it will reference both
  registers in the register pair. I.e. we'll have something like

    invoke-static {v0, v1} Lfoo;.bar(J)V

  convert it to `invoke-static {v0} ...` instead.

In this diff the naive register allocator only uses the range <-> no-range
conversion. I also tweaked it to try and convert fewer opcodes to the /range
form. Subsequent diffs will have the allocator make use of the canonize() API
as well.

Eventually I plan to have all IRInstructions converted to this canonical form
at load time, so that all our opts operate on that form.

Differential Revision: D4992947

fbshipit-source-id: f8a19371a9b10b38ce4e5a8c9d82e5c0ca88625d