Represent check-cast opcodes with a dest in the IR
Summary:
check-cast opcodes change the type of a register. As such, for
many analyses, it is semantically equivalent to creating a new value.
By representing the opcode in our IR as having a dest field, these
analyses can be simplified by not having to treat `check-cast` as a
special case.

Since `opcode::dests_size()` and friends are no longer accurate for
`check-cast` (since the actual encoded form doesn't have a dest but
the IRInstruction does), I've moved them into an `opcode_impl`
namespace to discourage accidental usage.

Differential Revision: D5027408

fbshipit-source-id: bc1b4c37901ee4eec6a307691bfa781085d67899