Inline constructor
Summary:
As title.

Until now we were only treating the builders with trivial constructors, in
which case we weren't inlining the call, just removing it for the case where the
builder gets removed.

  invoke-direct {v0}, Ljava/lang/Object;.<init>:()V
  return-void

In order to treat the ones more complex, we need to also inline the
constructor. While doing this we need to make sure we are:
* inlining constructor(s)
* removing the builder completly
* remove the super_class (which in this case we only restrict to Object) init calls

For the case where the inlining doesn't succeed or the removal, we need to
restore the method, since the verifier will complain if the constructor is not
called.

Other changes/moves:
* moved helper methods to RemoveBuildersHelper
* filter out builders that don't extend Object
*

Reviewed By: int3

Differential Revision: D4983353

fbshipit-source-id: 7a62ea892e1f95ebcd438b245e3bf51c90b206e6