`oatmeal`, a tool for dumping and building oatfiles.
Summary:
Skeleton of an oat-file manipulation tool, which will ultimately
be used to generate oat files quickly, on-device. Currently, it has
parsing code only, no building code. Starting with parsing was the best
way to learn about the oat format.

The parsing code includes a "memory accounter", which is useful for making
sure that the parsing is complete, i.e. that all pieces of the file are
understood to some degree by the code. This is critical for making sound
judgments about whether fast generation of the oat file is feasible.

There are a few small wheels re-invented in this code. This is intentional
because I want to keep the dependencies of this code as small as possible,
in order to keep the final binary size as small as possible, since this
will need to ship with the app.

Currently, only a single oat file version is supported. More versions will
be added as we go.

Also, this currently only reads the portion of the oat file that lives
in the .rodata section of the elf file. To read an oat file, you first need to do:

  $ objcopy -O binary --only-section=.rodata everything.oat everything.oat.rodata

And then use oatmeal on everything.oat.rodata

Reviewed By: dariorussi

Differential Revision: D4868982

fbshipit-source-id: 4fe526e62f9ad0a1ec06f3dcfd6697e53131177b