Implement instruction selection
Summary:
Added a `select_instructions` method that:

* Converts each `move-` and `const-` opcode to the smallest form that can address its operands
* Inserts move instructions as necessary for check-cast instructions that have different src and dest registers. Not-quite-obviously, the move opcode *must* come after the check-cast; see the code comments for why.

I also moved the logic that selects 2addr instructions from `to_dex_instruction` into the `select_instructions` function, where it seems to make more sense. It's now possible to have IRInstructions with 2addr-type opcodes -- we'll just throw an error if the src and dest are different when you try and convert it to a DexInstruction. This mirrors the approach used for check-cast and range-type opcodes.

In addition to shrinking our code slightly, this should also simplify some of our opts, especially the register allocator -- we can generate / choose the largest move instruction with the assurance that it gets shrunk down later.

I added a `next_release_gate` method to RedexContext in order to gate this change.

Reviewed By: justinjhendrick

Differential Revision: D5035274

fbshipit-source-id: 2ebe98926eb6c19269426ff010c36161e2181e09