Add option to provide a 'super class holder'
Summary:
Instead of assuming the builder constructor contains a call to the super class
constructor, provide a way to pass the class that will be initialized
in the constructor and that should be removed.

No change for the current builders, giving that we are not passing a super
class holder. To be used in the next change for the component builders.

NOTE: Instead of passing the super_class_holder, `remove_builder` could take
care of the super class init calls (and inline everything appropriatly and then
restore). For now, the callers have to make sure that the constructor only calls the `Object` constructor.

Not removing the builder in case the super_class != Object.

Reviewed By: justinjhendrick

Differential Revision: D5065555

fbshipit-source-id: b03eca467aa224ee3f2163307fc172da545be3a4