dce: rebuild CFG if we removed instructions
Summary:
Removing dead instructions can make it apparent that some basic blocks are
unreachable. Rebuild the CFG so we don't miss removing them.

In addition to the (tiny) size win, this simplifies register allocation -- right now
the allocator is very sensitive to dead code; it will throw if it cannot
determine the type of a dead register.

Differential Revision: D5091222

fbshipit-source-id: d2f6522c9336bea29122facbae0fda7b644b6156