Don't remove builders that escape the current method
Summary:
In case the builder escapes the method (it is passed to another function or
saved in a field or used in a conditional statement) not removing the builder.

Before removing the builder
* there shouldn't be any super classes method calls (all of those were inlined if we got to this point)
* there shouldn't be any super class fields used

We expect that the usage at this point is something like:

  // initialize builder
  // field setters + field usage
  // builder.release()

Note: Not enabling the pass yet, giving that we are not removing the builders.

Reviewed By: justinjhendrick

Differential Revision: D5065845

fbshipit-source-id: 5a152c1196ad166903f998a121b63133b7c437e8