slightly more conservative renamer
Summary:
While examining the code, I found some scary places we are using string literals containing class names that we are renaming. Some occur in very out of the way places like binder codegen for aidl. It's unclear if this is safe or dangerous, but I want to tighten up the renamer to play it safe. Instead of blacklist strings that occur in Class.forName("<literal>") - we will blacklist any class name we see in a const-string.

We give back 18k with this diff but we've got over a megabyte (8,000+ classes) getting renamed with graphql renaming in the near term, so I think we'll be ok.

Reviewed By: justinjhendrick

Differential Revision: D5068432

fbshipit-source-id: 63609ce107715661606214746e35daf1ca27e92b