add force_rename_hierarchies to renamer v2
Summary:
Add a new config to renamer to allow the renamer to be instructed to rename classes that fall under a hierarchy no matter what, and to also rewrite const-string opcodes that contain those class names (in external or internal format).

This is part of the larger set of work being done to support renaming of graphql models. graphql folks will at some point land a diff that contains codegen'd code that looks like:

    private static class Shard0 {
      private static String getClassName(int typeTag) {
        switch (typeTag) {
          case -2046891440:
            return "com.facebook.abtest.qe.protocol.sync.full.ViewerConfigurationQueryModels$ConfigurationParameterSetsConnectionModel$EdgesModel";
          case -274685910:
            return "com.facebook.account.switcher.settings.DBLRemoveRemoteDeviceMutationModels$DeviceBasedLoginRemoveRemoteDeviceMutationModel";
          case -1335458160:
            return "com.facebook.account.switcher.settings.FetchDBLNonceInfoQueryModels$FetchDBLNonceInfoQueryModel$DeviceBasedLoginNonceInfosModel$NodesModel";
    ...

When this graphql code is in place, these classes will be renameable from a graphql point of view, but Redex will not rename them because the class name appears in a string literal.  (see D5068432) With the change in this diff, we will add a config (see D5071922) that will enable renaming force renaming of  FragmentModels, and the bytecode will end up looking like

    private static class Shard0 {
      private static String getClassName(int typeTag) {
        switch (typeTag) {
          case -2046891440:
            return "x.A0z";
          case -274685910:
            return "X.9be";
          case -1335458160:
            return "X.0f8";
          case -316774380:
    ...

Reviewed By: justinjhendrick

Differential Revision: D5079119

fbshipit-source-id: 4e8e6fb2e8a260fff007f3553fa9f58f346e1b3a