integration with android_binary_for_automation
Summary:
Final piece of the native outliner. This wires it into the fbandroid build process, the test process, and sets up meaningful instrumentation/post verify tests.

The old world looked like this (not a dep graph, more of a conceptual flow):

    [code] => [pg/redex android_binary] => [other postprocess steps like hermes, optres, etc]

For a build that has the native outliner enabled, it will look like this (again, more of a flow than a buck dep graph:

    [code] => [pg/redex android_binary] => [collect outliner artifacts] => [codegen cpp] => [build so] => [glue so into apk] => [other postprocess steps like hermes, optres, etc]

This is /fairly straightforward except one piece; to glue the codegen'd .so into the chain of apk_genrules, we have to know the target architecture. We don't /actually/ know it directly in Buck... so we create a "dummy APK" which includes our .so's, that the real APK depends on. The cpu filter will do the right thing in placing the right arch .so's in the dummy APK, then it's just a matter of moving them from the dummy APK to the real APK. So really our dependency graph looks like this:

{F66828029}

Differential Revision: D4847509

fbshipit-source-id: 821a340cbfffc7936a389c17e242a3cf2c69f079