regalloc: Implement RegisterType
Summary:
Add a module that tracks the type of a register in order that we may insert the right move instructions when spilling it. We actually have a similar module right now called RegisterKind, but I'm replacing it with this one that is a slightly more accurate model with an explicit ZERO type and which can use both `src` and `dest` operands of an instruction to infer register types. It also uses ajv's BitVectorLattice for a more declarative and efficient implementation.

The old RegisterKind code is still around (though I stole a few of its unit tests), but will be removed in a separate diff along with the rest of the naive allocator.

Reviewed By: justinjhendrick

Differential Revision: D5094058

fbshipit-source-id: 5d26b2dd43ce5202b98729df06634947f2cc1f03