simpleinline: Properly take advantage of register allocator
Summary:
SimpleInline currently goes through a bunch of contortions to maximize register use while staying under the 16 register limit. Once the register allocator is live, this will no longer be necessary.

Additionally, there a number of call sites which we do not currently inline because the arguments are aliased. SimpleInline currently maps the parameter registers in the callee to the argument-passing registers in the caller, so aliasing results in different callee registers being mapped to the same caller register, creating havoc.

Under the new scheme,  we simply expand the caller register file by the size of the callee register file, and allocate the high registers to the callee. E.g. if we have a caller with registers_size of M and a callee with registers_size N, we will resize the caller's register file to M + N and map register k in the callee to M + k in the caller by inserting appropriate move instructions. The move instructions break aliases, so that is no longer an issue. The register allocator will later coalesce these registers where possible so we don't bloat the register file unnecessarily.

I ended up touching a bunch of code in order to avoid repeatedly looking up the FatMethod::iterator corresponding to the invoke instruction (which we need in order to insert the move instructions)... could have probably done it in a separate diff but it's a bit of a pain to split out at this point. The main piece of interesting code though is `gen_callee_reg_map_with_alloc` in Transform.cpp.

Differential Revision: D5098748

fbshipit-source-id: af5f453e509343509d3b7ca19e0e4815bab25e71