Introduce global assume_regalloc config, lift 16reg limit for peephole
Summary:
It seems that the fix in {D4784032} misses handling the case where the replacement pattern for the with-move-result case has a lower register size limit than the no-move-result case. In that case the with-move-result case may fail to match even though the output register is live; we then fall through to the no-move-result case and generate the wrong peephole replacement. I've added a flag -- `assume_regalloc` -- to skip the register size limit check altogether once the register allocator is turned on.

This does mean that there are latent bugs in peephole even without the register allocator; I suspect we can fix them simply by replacing the move opcodes in the pattern replacements to use the larger move/16 opcodes, but I haven't gone to the trouble of testing it since regalloc should land very soon.

I chose to require `assume_regalloc` to be set in the config file instead of having RegAllocPass set it automatically because 1) it's useful to test the register allocation pass without changing the current behavior of the other passes and 2) it's pretty weird for a pass to modify a global config setting.

Reviewed By: justinjhendrick

Differential Revision: D5102469

fbshipit-source-id: 2a2e2d5794ba9445c47e7e2a2db788ae09d20f4a