Allow separate specification of dex file pathnames and locations
Summary:
Allow separate specification of dex file pathnames and locations
to be stored in the oat file. This is useful for tests (so that the generated
file doesn't have tmp pathnames in it) and also for comparing against
known-good files from dex2oat (by allowing us to match the file listing exactly).

Reviewed By: dariorussi

Differential Revision: D5041952

fbshipit-source-id: 7209da1bacc960ea457d19bee18112c0cc860dc7