fix make_unique
Summary:
redex has its own version of make_unique that isn't correct: it creates a temporary copy of all the arguments passed in. At best this just results in an unnecessary copy, at worst it results in incorrect behavior or a crash if someone isn't actually expecting a copy (which is how I found this issue). Fix it by making make_unique forward args as rvalues as expected.

Not sure who to add on this review, so I'm just going to add some people I've worked with recently and recent committers to this file.

Reviewed By: dariorussi

Differential Revision: D5129179

fbshipit-source-id: 8b23905f066242a0eb61d929ba91346e1299d87b