Fix builder transfer reach
Summary:
When checking for which registers hold a builder, we were not taking into
account the case where a register that previously was holding a builder is
overwritten by a new instance of a different type.

What happened was that for example for

  // v1 holds a builder instance
  new-instance v1, 'DifferentType'
  // at this point v1 no longer holds buidler type

after the new-instance we weren't taking the change into account.

Reviewed By: int3, newobj

Differential Revision: D5123211

fbshipit-source-id: e77e6126bec83e006100ee14dcbdeca79e9520ad