Don't inline methods that contain cross store references
Summary:
we introduce a concpet of store or virtual store for the inliner. The primary dex is in one store the remaining dexes in the first store are in another store and every other store is in its own store.
Then when inlining if a method contains a cross store reference we leave it untouched.
That should take care of the primary DEX and of any other cross store problem.
I have verified deps_only.config that the references are untouched.
We lose a bunch of method and I am going to verify manually the reason for that.
Overall though things look good and the APK with method with throw inlined loads.

build[automation_fbandroid_redexunstable,automation_messenger_redexunstable]
redexstats[fb4a_unstable,messenger_unstable]

Differential Revision: D5100020

fbshipit-source-id: ed43f75c19aaec938afbb4c13f4f4374030daf17